+++ examples/standard-use/vertex-ai-stream-with-tools.ts
@@ -1,0 +1,34 @@
+#!/usr/bin/env ts-node
+
+import { config } from "dotenv";
+import { JorEl } from "../../src";
+import { getWeather } from "../_utilities/get-weather";
+import { z } from "zod";
+
+config({ path: "../../.env" });
+
+const main = async () => {
+  // Create instance
+  const jorEl = new JorEl({ vertexAi: true });
+
+  // Will return a stream of strings
+  const stream = jorEl.stream("What is the weather in Sydney?", {
+    tools: [
+      {
+        name: "get_weather",
+        description: "Get the current temperature and conditions for a city",
+        executor: getWeather,
+        params: z.object({ city: z.string() }),
+      },
+    ],
+  });
+
+  // Print each chunk
+  for await (const chunk of stream) {
+    process.stdout.write(chunk);
+  }
+
+  process.stdout.write("\n");
+};
+
+void main();